feat: add exercises section component
